Player Respawns on Spike Trigger Enter
There is an empty game object with a box collider that is a trigger with the tag of respawn point. The height of the invisible box goes to the height of the tilemap and the width is narrow. On trigger enter the player stores the location if the respawn point and uses that location when the player falls in the spikes.